@@ -1,11 +1,6 @@
-import sys
+from typing import Annotated
 
 import marshmallow.fields
-
-if sys.version_info >= (3, 9):
-    from typing import Annotated
-else:
-    from typing_extensions import Annotated
 
 Url = Annotated[str, marshmallow.fields.Url]
 Email = Annotated[str, marshmallow.fields.Email]
