@@ -1,15 +1,9 @@
+import dataclasses
 import inspect
 import typing
 import unittest
-from typing import TYPE_CHECKING, Any, cast
+from typing import TYPE_CHECKING, Any, Final, Literal, cast
 from uuid import UUID
-
-try:
-    from typing import Final, Literal  # type: ignore[attr-defined]
-except ImportError:
-    from typing_extensions import Final, Literal  # type: ignore[assignment]
-
-import dataclasses
 
 from marshmallow import Schema, ValidationError
 from marshmallow.fields import UUID as UUIDField
@@ -17,7 +11,7 @@ from marshmallow.fields import Field, Integer
 from marshmallow.fields import List as ListField
 from marshmallow.validate import Validator
 
-from marshmallow_dataclass import NewType, class_schema
+from marshmallow_dataclass2 import NewType, class_schema
 
 
 class TestClassSchema(unittest.TestCase):
@@ -126,6 +120,34 @@ class TestClassSchema(unittest.TestCase):
         self.assertIsInstance(schema.fields["uuid"], UUIDField)
         self.assertIsInstance(schema.fields["n"], Integer)
 
+    def test_use_type_mapping_from_base_schema_alias(self):
+        """Identical to test_use_type_mapping_from_base_schema, but using `list` instead of `typing.List`"""
+
+        class CustomType:
+            pass
+
+        class CustomField(Field):
+            pass
+
+        class CustomListField(ListField):
+            pass
+
+        class BaseSchema(Schema):
+            TYPE_MAPPING = {CustomType: CustomField, list: CustomListField}
+
+        @dataclasses.dataclass
+        class WithCustomField:
+            custom: CustomType
+            custom_list: list[float]
+            uuid: UUID
+            n: int
+
+        schema = class_schema(WithCustomField, base_schema=BaseSchema)()
+        self.assertIsInstance(schema.fields["custom"], CustomField)
+        self.assertIsInstance(schema.fields["custom_list"], CustomListField)
+        self.assertIsInstance(schema.fields["uuid"], UUIDField)
+        self.assertIsInstance(schema.fields["n"], Integer)
+
     def test_filtering_list_schema(self):
         class FilteringListField(ListField):
             def __init__(
@@ -154,6 +176,49 @@ class TestClassSchema(unittest.TestCase):
                 metadata={"max": 2.5, "min": 1}
             )
             constrained_strings: typing.List[str] = dataclasses.field(
+                metadata={"max": "x", "min": "a"}
+            )
+
+        schema = class_schema(WithCustomField, base_schema=BaseSchema)()
+        actual = schema.load(
+            {
+                "constrained_floats": [0, 1, 2, 3],
+                "constrained_strings": ["z", "a", "b", "c", ""],
+            }
+        )
+        self.assertEqual(
+            actual,
+            WithCustomField(
+                constrained_floats=[1.0, 2.0], constrained_strings=["a", "b", "c"]
+            ),
+        )
+
+    def test_filtering_list_schema_alias(self):
+        class FilteringListField(ListField):
+            def __init__(
+                self,
+                cls_or_instance: typing.Union[Field, type],
+                min: typing.Any,
+                max: typing.Any,
+                **kwargs,
+            ):
+                super().__init__(cls_or_instance, **kwargs)
+                self.min = min
+                self.max = max
+
+            def _deserialize(self, value, attr, data, **kwargs) -> list[typing.Any]:
+                loaded = super()._deserialize(value, attr, data, **kwargs)
+                return [x for x in loaded if self.min <= x <= self.max]
+
+        class BaseSchema(Schema):
+            TYPE_MAPPING = {list: FilteringListField}
+
+        @dataclasses.dataclass
+        class WithCustomField:
+            constrained_floats: list[float] = dataclasses.field(
+                metadata={"max": 2.5, "min": 1}
+            )
+            constrained_strings: list[str] = dataclasses.field(
                 metadata={"max": "x", "min": "a"}
             )
 
