@@ -1,9 +1,9 @@
 import doctest
 
-import marshmallow_dataclass
+import marshmallow_dataclass2
 
 
 def load_tests(_loader, tests, _ignore):
     # Load all the doctests defined in the module
-    tests.addTests(doctest.DocTestSuite(marshmallow_dataclass))
+    tests.addTests(doctest.DocTestSuite(marshmallow_dataclass2))
     return tests
